requirements.txt
google-generativeai==0.2.1
pandas
tenacity